Remove Error inheritance
This turns out to break in node5 (it didn't in node 0.11.13, where we were
using this lib).

Basically, inheriting from Error overwrites the `this.stack` field and
does not let anything else write to it. This means that the "original"
(inner, unmodified) stacktrace will still be gotten from the error, which
is one of the main things we're trying to avoid with this lib.

Unfortunately this means that `someCubeError instanceof Error` will no
longer work. Should be a major bump.